small refactor for better mvc principle application
